Fixed formatting and class naming
