+++ src/main/java/pl/jsolve/sweetener/mapper/builder/strategies/CollectionElementsMapperStrategy.java
@@ -1,18 +1,18 @@
-package pl.jsolve.sweetener.mapper.custom.strategies;
+package pl.jsolve.sweetener.mapper.builder.strategies;
 
 import java.lang.reflect.Array;
 import java.util.Collection;
 
-import pl.jsolve.sweetener.mapper.custom.CustomMapper;
+import pl.jsolve.sweetener.mapper.builder.MapperBuilder;
 
 public class CollectionElementsMapperStrategy implements CustomMapperStrategy {
 
 	private final Class<?> elementsType;
-	private final CustomMapper mapper;
+	private final MapperBuilder mapper;
 
 	public CollectionElementsMapperStrategy(Class<?> elementsType) {
 		this.elementsType = elementsType;
-		mapper = CustomMapper.toType(elementsType).usingAnnotations().usingTypeConvertion();
+		mapper = MapperBuilder.toType(elementsType).usingAnnotations().usingTypeConvertion();
 	}
 
 	@Override
